refactor: consolidate AttachmentStore and TranscriptCacheStore into JsonFileStore<T>
